+++ crates/sevenmark-html/examples/render.rs
@@ -19,6 +19,7 @@ fn main() {
     let render_start = Instant::now();
     let config = RenderConfig {
         edit_url: Some("/edit/TestDocument"),
+        file_base_url: Some("https://cdn.example.com/"),
         document_base_url: Some("/Document/"),
         category_base_url: Some("/Category/"),
     };